Move column creation to BDR 1.0.3.0
For dump/restore across BDR versions we need all columns
to be the same in the BDR tables. So add the columns to 1.0.3,
even though they'll be unused in 1.0.x.

This requirement should go away later if we start excluding extension
catalog tables from dumps, and instead syncing them with COPY. Which
would be better anyway...