Don't manually advance replication origins
There is no need to call replorigin_session_advance during
process_remote_commit, and doing so has no useful effect.

replorigin_session_origin_lsn is already set to the remote commit's start LSN,
and that's what we report throughout BDR. It gets flushed to the commit record
when we commit locally.

Calling replorigin_session_advance(...) with the commit end+1 pointer wasn't
safe; it didn't emit WAL, so if we crashed its effects would be lost and the
replication origin would be set back to the commit start pointer during redo.

(Note that the immediately following commit will change
replorigin_session_origin_lsn to point to the end-of-commit+1 lsn, not
commit start LSN).