Use end-of-commit LSN in progress tracking
We previously used the start-of-commit-record LSN for replication origin
progress tracking; i.e. if we just finished applying a commit that started at
1/1, we'd store 1/1 in the replication origin. However, START_REPLICATION's
input LSN is inclusive, if you ask for 1/1 and there's a commit starting
exactly at 1/1, you'll get that commit replayed again.

This didn't happen much in practice because standby status updates tended to
promptly advance confirmed_flush, so the requested start position would be
skipped over in favour of starting at confirmed_flush.

The effect of the problem was that duplicate xacts could replay after crash if
the other end had not yet received feedback for some confirmed LSN past
commit-start or had been restarted before flushing that feedback to disk. (It
doesn't dirty the slot for feedback, so there's a large window there).

To fix this, keep track of commits using the end-of-commit LSN, not the
start-of-commit LSN.

When sending START_REPLICATION increment the tracked position by 1 so we're
always asking to start from the beginning of the next record. This isn't
strictly necessary since asking for any LSN after the start-of-commit will skip
a given commit, but it provides a safeguard when we're replaying from a BDR 1.0
node that doesn't yet have this fix, and it's cleaner to request what we
actually want.

Since there's only room for one LSN in the BEGIN message, substitute the
end-of-commit LSN for the start-of-commit LSN there.

This is compatible with BDR 1.0. When processing messages from BDR 1.0 we'll
still store the start-of-commit LSN in our replication origin, but we'll
increment it by 1 when requesting replay start so the bug won't happen. When
BDR 1.0 replays our messages it'll store end-of-lsn in its replication origin,
which will also be fine, and will defend 1.0 instances replaying from 2.0
instances against the bug. (Until/unless this is backpatched, at which point
it stops mattering).