Fix SPI state clobbering in bdr_maintain_db_workers()
While scanning for slots to remove we could clobber SPI state after
the first iteration in bdr_maintain_db_workers(). This issue looks
to be new as of the node name output but it's worth verifying that
and checking if backport is needed.

Symptom is failure of nodes after 1st parted node to properly drop slots and
connections after being parted.