bdr_init_copy: ddl lock retry loop
When bdr_init_copy acquired the global DDL lock to INSERT the bdr.bdr_nodes
entry for the newly joining node, it did so without considering that the
lock acquisition might fail. The DDL lock is a non-blocking (nowait) lock
(2ndQuadrant/bdr-private#59) so we can't just fire and forget.

Also, bdr_init_copy was taking the global DDL lock, but it wasn't holding it
for the duration of the following INSERT, so it actually had limited effect.

Fix it by implementing a retry and backoff loop - one that also handles
deadlocks, connection loss, etc. We'll now retry if we fail to get the lock.

Fixes 2ndQuadrant/bdr-private#58
Fixes 2ndQuadrant/bdr-private#64