Fix up logic in offline join test
The offline join test was assuming that bdr.bdr_group_join would block, which
it won't. It also expected a DDL lock to hang around after the xact that acquired
it exited and some other issues.

Test still doesn't pass, now it shows an actual problem in logical join.

Related to #69