Add a node-local 'lock_state' to global locks
Global locking now keeps track of a lock phase/state that exists
separately to the progress tally counters, lockcount, and
this_xact_acquired_lock field. This allows other backends on a node
to query the lock state, determining

* Whether the local node is the acquiring node
* Currently active locking phase
* Number of confirmations/acknowledgements pending

etc.

This can be exposed to users for monitoring purposes
(2ndQuadrant/bdr-private#60).

Prior to this change the only backend that knew which backend was acquiring the
lock and whether it was fully acquired yet was the backend that requested the
lock.

A bit of refactoring was required to get this to work and there are a lot of
added assertions, so there's some limited stability risk here. But the test
suite is happy. The assertions won't affect production builds but the extra
LWLock acquisitions could. However they're only really an issue when you have
multiple BDR nodes in one Pg instance, which is really only done in testing/QA.

The new state field is only informational so far and won't control any locking
behavour.

It can potentially be extended to allow other local backends to continue to do
DML while the node acquires and holds the DDL lock (at the risk of delaying the
DDL). Local backends can then rely on Pg's normal locking. But so far it isn't
used for that.