Acquirer/holder of global lock is now excempt from write lock
The node that acquires the global DDL lock in 'write' mode is now exempt from
the write-lock DML lockout. Local write transactions are not cancelled, and
new local write transactions are not prohibited.

This could cause DDL on the executing node to take longer if it has to
wait on locks held by other local xacts. But that's the same as stock
PostgreSQL and the admin can deal with it if required. We can allow local
xact kill after a timeout if this proves to be a problem in practice.

The upside is that in single-master fast-failover deployments, which is
what lots of people seem to actually want to use BDR for, it's much less
distruptive to the "primary" node.

Fixes 2ndQuadrant/bdr-private#78