Fix possible deadlock in find_pkey_tuple
When retrying because of concurrency, the original index scan remains
open and could keep a pin on the buffer.  Trying to open another index
scan without releasing the first could lead to deadlocking on the buffer
lock.

So far, this is an untested fix, because of bugs in the makefiles
prevent me from compiling BDR at the moment.  We'll see what the build
systems have to say.