Reset MessageContext on commit to limit possible growth
This isn't perfect, because we can potentally process hundreds of millions of
messages in a single big txn, but it's a help.

I don't want to add a counter-based reset just yet because I'm concerned that
we won't hit it during routine testing, and it'll obscure accesses to free'd
memory that only happen during very big txn processing. We need to improve big
tx test coverage before this can be done.